==> 04_TransformationDay2.Rmd <==
---
editor_options: 
  markdown: 
    wrap: 72
---

# Data Transformation Day 2

> "Every day, three times per second, we produce the equivalent of the
> amount of data that the Library of Congress has in its entire print
> collection, right? But most of it is like cat videos on YouTube or
> 13-year-olds exchanging text messages about the next Twilight movie."
>
> ::: {style="text-align: right"}
> --- Nate Silver
> :::

First off...

<h4>Advanced R Challenges</h4>

Explain how the order of factors changed within the plots for dat2. How
do facet_grid and facet_wrap work?

```{r advancedR, eval=FALSE}
#Can you reorder
library(ggplot2)
library(dplyr)
library(forcats)
dat2 <- structure(list(School = structure(c(2L, 2L, 2L, 1L, 1L, 1L, 3L, 
                                            3L, 3L), .Label = c("School1", "School2", "School3"), class = "factor"), 
                       Year = c(2015L, 2014L, 2013L, 2015L, 2014L, 2013L, 2015L, 
                                2014L, 2013L), Rate = c(80L, 90L, 11L, 70L, 50L, 30L, 60L, 
                                                        50L, 40L), Order = c(1L, 1L, 1L, 2L, 2L, 2L, 3L, 3L, 3L), 
                       Order2 = c(1L, 2L, 3L, 1L, 2L, 3L, 1L, 2L, 3L), 
                       Order3 = structure(c(NA_integer_,
                                            NA_integer_, NA_integer_, NA_integer_, NA_integer_, NA_integer_, 
                                            NA_integer_, NA_integer_, NA_integer_), 
                                          .Label = c("School2","School1","School3"), class = "factor")),
                  .Names = c("School","Year", "Rate", "Order", "Order2", "Order3"), 
                  row.names = c(NA,-9L), class = c("tbl_df", "tbl", "data.frame")) %>%
  mutate(School = fct_relevel(School, c("School2","School1","School3")))



p <- ggplot(dat2, aes(x=reorder(Year,Order2), Rate), labels = c("School1","School2","School3")) +
  geom_bar(stat = "identity", width = 0.5)
p + facet_grid(. ~School)


dat2 %>%
  mutate(School = fct_relevel(School, c("School2","School1","School3"))) %>%
  ggplot(aes(x=reorder(Year,Order2), Rate), labels = c("School1","School2","School3")) +
  geom_bar(stat = "identity", width = 0.5) +
  facet_grid(. ~School)
```

## Day 2 {.unnumbered}

```{r setup, echo=FALSE,include=FALSE}
# install.packages("forcats") # You might need this because it doesn't load with the tidyverse.
rm(list = ls())
require(tidyverse)
require(forcats)
```

*Note*: Much of this lesson can be found in the "factors" chapter of R
for Data Science by Hadley Wickham: <http://r4ds.had.co.nz/factors.html>

```{r eval=FALSE}
?gss_cat #See more about the variables in our dataset

View(gss_cat)
```

```{r gss}
gss_cat %>%
  count(race) #We can see counts of each race or religion

gss_cat %>%
  count(relig) %>%
  arrange(desc(n)) #We can arrange them in descending order on the count column (n)

gss_cat %>%
  count(age) #Or we can count by age...


gss_cat %>%
  count(race) %>%
  mutate(freq = (n / sum(n))*100) #We can see relative proportions in the sample

levels(gss_cat$rincome)
#Releveling factors with forcats
df <- forcats::gss_cat
```

```{r ggplotintro}
#Remember, none of these outcomes are being saved into any objects with <- 
#This is just for visualizing and exploring the data.

#We can visualize this information with ggplot2
require(ggplot2) #We will get more into ggplot next week, 
#but for now here is an introduction.

ggplot(data = gss_cat, mapping = aes(x =race)) +
  geom_bar()


# By default, ggplot2 will drop levels that don’t have any values. 
# You can force them to display with:
ggplot(gss_cat, aes(race)) +
  geom_bar() +
  scale_x_discrete(drop = FALSE)

```

When working with factors, the two most common operations are: 1.
Changing the order of the levels 2. Changing the values of the levels.

Why would you want to do this?

It's often useful to change the order of the factor levels in a
visualisation.

Let's get the average number of hours spent watching TV per day across
religions:

Note: if you put () on the outside of your assignment, it will also
print

```{r relig_summary}
(relig_summary <- df %>% #Get your data.frame and pipe into the columns
  group_by(relig) %>% #group by religion so that each summary will be for one religion
  summarise( #Get a summary for the following transformations
    age = mean(age, na.rm = T), #Get the mean of age and name it "age"
    tvhours = mean(tvhours, na.rm = TRUE), #Get the mean of tv hours and name it "tvhours"
    n = n() #Count the number of observations in each group
  ))
```

Because there were NA's and we wanted to do math we needed to remove
them with: na.rm = TRUE

## Change the order of the levels

Plot with ggplot, arguments are ggplot(data, mapping) + geom_type()
mapping is usually paired with the aes() function, which refers to
aesthetics arguments to aes are usually aes(x, y, group)

```{r plotrelig}
ggplot(data = relig_summary, 
       mapping = aes(y = relig, x =tvhours)) + 
  geom_point()

```

Interesting, but we could pack more info into this visualization

```{r plotmore}
ggplot(relig_summary, aes(x = tvhours, y = fct_reorder(.f = relig, .x = tvhours))) +
  geom_point()

```

Now we have reordered religion by tv hours watched.

The key argument is fct_reorder("factor to order", "column to be ordered
by")

We can do it in descending order as well:

```{r plotdescend}
ggplot(relig_summary, aes(tvhours, fct_reorder(relig, desc(tvhours)))) +
  geom_point() 
```

We just put desc() around tvhours

As you start making more complicated transformations, I'd recommend
moving them out of aes() and into a separate mutate() step. For example,
you could rewrite the plot above as:

```{r fct_reorder}
relig_summary %>%
  mutate(relig = fct_reorder(relig, tvhours)) %>%
  ggplot(aes(tvhours, relig)) +
  geom_point()

mtv <- relig_summary %>%
  summarise(mtv = mean(tvhours))

relig_summary %>% ggplot(aes(x = tvhours, y = relig, color = relig)
) + geom_point() + geom_vline(aes(xintercept = mtv$mtv)) 

levels(relig_summary$relig)

(testlevel <- relig_summary %>%
  mutate(relig = fct_reorder(relig, tvhours)))

levels(testlevel$relig) #The levels have changed even if the visualization did not.
```

This is most important for plotting

Let's look at how average age varies across reported income level:

```{r plotcount}
(rincome_summary <- df %>% #Save your df summary as something descriptive
  group_by(rincome) %>% #Group by income so that all summaries will be for those levels
  summarise( #Create summaries for the following:
    age = mean(age, na.rm = TRUE), #Mean age with Na's removed named "age"
    tvhours = mean(tvhours, na.rm = TRUE), #Mean tvhours named "tvhours"
    n = n() #Count number of observations in each column
  ))

ggplot(rincome_summary, aes(age, fct_reorder(rincome, age))) + 
  geom_point()
```

Reported income is an ordinal variable so it should be reported that way
However fct_reorder will often create an arbitrary order by the second
argument (age in this case).

Use fct_relevel to have an ordered outcome:

```{r fct_relevel ordered}
ggplot(rincome_summary, aes(age, fct_relevel(rincome))) +
  geom_point()
```

Now "Not applicable" is up at the top away from the other "special
responses" Let's fix that:

```{r plotNAs}
ggplot(rincome_summary, aes(age, fct_relevel(rincome, "Not applicable"))) +
  geom_point()
```

We put it at the end! Now age and income are ordered! However, larger
numbers map to the top (Larger is higher in conceptual metaphor theory)

```{r plotfixed, eval=FALSE}
ggplot(rincome_summary, aes(age,x = fct_relevel(desc(rincome), "Not applicable"))) +
  geom_point() 

```

This doesn't work in the aes() function, so we have to do it earlier!

```{r plotfull}
ordered_rincome <-rincome_summary %>%
  mutate(rincome = fct_relevel(rincome, "Not applicable", after = 0))

levels(ordered_rincome$rincome) #not applicable has moved...

ggplot(ordered_rincome, aes(age, fct_relevel(rincome, "Not applicable"))) +
  geom_point() #But still plotting int he wrong order

ordered_rincome <- rincome_summary %>%
  mutate(rincome = fct_relevel(rincome, "Not applicable", after = 0)) %>%
  mutate(rincome = fct_rev(rincome)) %>% #Now we can use fct_rev()
  filter(rincome != "Not applicable",
         rincome != "No answer")

ggplot(ordered_rincome, aes(age,rincome)) +
  geom_point()
```

Let's do some more reordering:

```{r plotage}
by_age <- data.frame(df) %>% #Get data.frame
  na.omit() %>% #Get rid of Na's in age to do maths
  count(age, marital) %>% # count observations by age and marital status
  group_by(age) %>% #Group by age for the following summary
  mutate(prop = n / sum(n)) #Get proportion of counts per age

ggplot(by_age, aes(age, prop, colour = marital)) +
  geom_line(na.rm = TRUE)

ggplot(by_age, aes(age, prop, colour = fct_reorder2(marital, age, prop))) +
  geom_line() +
  labs(colour = "marital")
```

## RcolorBrewer

Let's get researched colors for proper visualization best practices
from: Brewer, Cynthia A., 200x. <http://www.ColorBrewer.org>, accessed
09/27/2018.

```{r colorBrewer}
library(RColorBrewer)


brewer.pal.info

pal1 <- brewer.pal(n = 6, "Dark2") #Try different palettes


display.brewer.all()

ggplot(by_age, aes(age, prop, colour = fct_reorder2(marital, age, prop))) +
  scale_color_brewer(palette = "Dark2") +
  geom_line() + theme_classic() +
  labs(colour = "marital")

  
ggplot(by_age, aes(age, prop, colour = fct_reorder2(marital, age, prop))) +
  geom_line() +
  labs(colour = "marital") + theme_dark() +
  scale_color_brewer(palette = "Accent")

ggplot(by_age, aes(age, prop, colour = fct_reorder2(marital, age, prop))) +
  geom_line() +
  labs(colour = "marital") + theme_void() +
  scale_color_brewer(type = "qual")
```

Finally, for bar plots, you can use <code>fct_infreq()</code> to order
levels in increasing frequency: this is the simplest type of reordering
because it doesn't need any extra variables. You may want to combine
with <code>fct_rev()</code>.

```{r marital}
df %>%
  mutate(marital = marital %>% fct_infreq() %>% fct_rev()) %>% 
  ggplot(aes(marital)) +
  geom_bar()

```

Notice that you can pipe these arguments WITHIN the mutate() function!

**Challenge 1**

For each factor in df identify whether the order of the levels is
arbitrary or principled.

```{r challenge1}
rev(levels(df$marital))

ggplot(df, aes(year,fill = relig)) + 
  geom_histogram(bins = 8) +
  scale_fill_brewer(palette = "Set3")

ggplot(df, aes(year,fill = fct_lump(relig,n = 8))) + 
  geom_histogram(bins = 8) +
  scale_fill_brewer(palette = "Set3")
```

How would you determine this?

## Modifying Factor Levels

Here's a count of the levels in party_id

```{r count}
df %>% count(partyid)
```

<code>fct_recode</code> let's us rename variables (new varname left =
old varname right)

```{r fct_recodeCount}
df %>%
  mutate(partyid = fct_recode(partyid,
                              "Republican, strong"    = "Strong republican",
                              "Republican, weak"      = "Not str republican",
                              "Independent, near rep" = "Ind,near rep",
                              "Independent, near dem" = "Ind,near dem",
                              "Democrat, weak"        = "Not str democrat",
                              "Democrat, strong"      = "Strong democrat"
  )) %>%
  count(partyid) #Count party ID

```

<code>fct_recode()</code> will leave levels that aren't explicitly
mentioned as is, and will warn you if you accidentally refer to a level
that doesn't exist.

If you want to collapse a lot of levels, <code>fct_collapse()</code> is
a useful variant of <code>fct_recode()</code>. For each new variable,
you can provide a vector of old levels:

```{r fct_collapseCount}
df %>%
  mutate(partyid = fct_collapse(partyid,
                                other = c("No answer", "Don't know", "Other party"),
                                rep = c("Strong republican", "Not str republican"),
                                ind = c("Ind,near rep", "Independent", "Ind,near dem"),
                                dem = c("Not str democrat", "Strong democrat")
  )) %>%
  count(partyid)
```

Pretty cool huh?! We just created in groupings to plot or do analyses
on.

Sometimes you just want to lump together all the small groups to make a
plot or table simpler. That's the job of <code>fct_lump()</code>

```{r lumping}
df %>%
  count(relig) #This is before lumping

df %>%
  mutate(marital = fct_lump(marital)) %>%
  count(marital)

df %>%
  mutate(relig = fct_lump(relig)) %>%
  count(relig) #This is default after lumping

levels(df$relig)
```

It looks like Protestant had the largest count, so by default it grouped
everything else and Protestant together.

You can collapse across specified lumps...

```{r fct_lump}
df %>%
  mutate(relig = fct_lump(relig, n = 10)) %>% #Specify lumps here
  count(relig, sort = TRUE)

```

## Play time

Answer some interesting questions about this dataset

```{r df}
df
```

What is the proportion of race in each religion?

```{r proportion, eval=FALSE}
df %>%
  group_by(relig) %>%
  count(...) %>%
  mutate(prop = ...)
```

Try to plot it

```{r tryplot, eval=FALSE}
df %>%
  group_by(relig) %>%
  count(race) %>%
  mutate(prop = n/sum(n)) %>%
  ggplot(aes(relig, ..., fill = ...)) + 
  geom_bar(stat = "...", position = "...") +
  theme(...)
```

## Solutions

Play 1

```{r play1}
View(df %>%
  group_by(relig) %>%
  count(race) %>%
  mutate(prop = n/sum(n)))
```

Play 2

```{r play2}
df %>%
  group_by(relig) %>%
  count(race) %>%
  mutate(prop = n/sum(n)) %>%
  ggplot(mapping = aes(relig, prop, fill = relig)) + 
  geom_bar(stat = "identity", position = "dodge") +
  theme(axis.text.x = element_text(angle = 90, hjust = 1))

df %>%
  group_by(relig) %>%
  count(race) %>%
  mutate(prop = n/sum(n)) %>%
  ggplot(mapping = aes(x = relig, y = prop, color = race)) +
  geom_line(aes(group = race)) + geom_point() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))


df %>%
  count(relig, race) %>%
  group_by(relig) %>%
  mutate(prop = n/sum(n))
```
