==> 05_Visualize.Rmd <==
```{r include=FALSE, cache=FALSE}
set.seed(1014)
options(digits = 3)

knitr::opts_chunk$set(
  comment = "#>",
  collapse = TRUE,
  cache = TRUE,
  out.width = "70%",
  fig.align = 'center',
  fig.width = 6,
  fig.asp = 0.618,  # 1 / phi
  fig.show = "hold"
)

options(dplyr.print_min = 6, dplyr.print_max = 6)
```
# Visualization with ggplot2
>“...few people will appreciate the music if I just show them the notes. Most of us need to listen to the music to understand how beautiful it is. But often that's how we present statistics; we just show the notes we don't play the music.” 
><div style="text-align: right">
> --- Hans Rosling </div>

>“At their best, visualizations of data effortlessly reveal the true nature of information.” 
><div style="text-align: right">
> --- Lace Padilla </div>

First off...

 <h4>Advanced R Challenges</h4>
 
Create a custom theme for your graphs!
```{r advancedR}
require(tidyverse)
#Example 1.
my.axis.font<-theme(axis.title.x = element_text(size=18), axis.title.y = element_text(size=18),
                    axis.text.x = element_text(size=18), axis.text.y = element_text(size=18), 
                    plot.title=element_text(size=18,face="bold", hjust = .5),
                    legend.title=element_text(size=14),legend.text=element_text(size=14))

# theme_mine <- theme(...)
# example 2.
### XKCD theme
theme_xkcd2 <- theme(
  panel.background = element_rect(fill="white"), 
  axis.ticks = element_line(colour=NA),
  panel.grid = element_line(colour="white"),
  axis.text.y = element_text(colour=NA), 
  axis.text.x = element_text(colour="black"),
  text = element_text(size=16, family="Humor Sans")
)

###Jittery line###
xkcd_line <- function(x, y, color) {
  len <- length(x);
  rg <- par("usr");
  yjitter <- (rg[4] - rg[3]) / 1000;
  xjitter <- (rg[2] - rg[1]) / 1000;
  x_mod <- x + rnorm(len) * xjitter;
  y_mod <- y + rnorm(len) * yjitter;
  lines(x_mod, y_mod, col='white', lwd=10);
  lines(x_mod, y_mod, col=color, lwd=5);
}

###Off-axis###
xkcd_axis <- function() {
  rg <- par("usr");
  yaxis <- 1:100 / 100 * (rg[4] - rg[3]) + rg[3];
  xaxis <- 1:100 / 100 * (rg[2] - rg[1]) + rg[1];
  xkcd_line(1:100 * 0 + rg[1] + (rg[2]-rg[1])/100, yaxis,'black')
  xkcd_line(xaxis, 1:100 * 0 + rg[3] + (rg[4]-rg[3])/100, 'black')
}

#Now use it to  plot some graphs.
mpg %>%
  ggplot() + geom_point(aes(x = displ, y = hwy, color = class)) +
  labs(title = "My Plot") +
  my.axis.font

#Some squiggly lines:
d <- data.frame(x=1:100)
d$one <- exp(-((d$x - 50)/10)^2)
d$two <- sin(d$x/10)
plot.new()
plot.window(
  c(min(d$x),max(d$x)),
  c(min(c(d$one,d$two)),max(c(d$one,d$two))))
xkcd_axis()
xkcd_line(d$x, d$one, 'red')
xkcd_line(d$x, d$two, 'blue')
```

## Steps of Visualization {-}

Step 1: Get some data

```{r setup, include=FALSE, message=FALSE}
rm(list = ls()) #clear environment
require(tidyverse) #You can use require() or library() to load the package
require(kableExtra)
```
<code>ggplot2</code> has the mpg dataset.
```{r mpg}
kable(ggplot2::mpg[1,1:11])
```
Let's save it to something...
```{r saveit}
df <- ggplot2::mpg
```
take a look at it...
```{r, eval=FALSE}
View(df)
```
Also look at its dimensions up in the environment ---->
(usually it's in the upper-right corner)
or print them to the console...
```{r dims}
dim(df)
str(df) #what kind of classes are we dealing with?
```
Where can I read more about this dataset?


hint:
```{r questions}
?mpg
```
You can plot this data quickly in many ways:
```{r, ezplots}
stem(mpg$displ)
hist(mpg$displ)
plot(mpg$displ)
plot(mpg$cty, mpg$hwy)
```
However, the ggplot package offers a structure and flexibility to create
almost any plot you can imagine. Other packages work with ggplot to
create the rest.

## ggplot2
Remember this scaffolding for your ggplots:
```{r scaffolding, eval=FALSE}
ggplot(data = <DATA>) +
  <GEOM_FUNCTION>(
    mapping = aes(<MAPPINGS>),
    stat = <STAT>,
    position = <POSITION>
  ) +
  <COORDINATE_FUNCTION> +
  <FACET_FUNCTION>
```
ggplot consists of a data.frame (or a tibble), a coordinate system,
and geoms (marks that represent data values)

This is usually how we start:
```{r baseggplot}
df %>% #the tibble
  ggplot(mapping = aes(x = displ, y = hwy)) + #the x and y coordinates (what goes on the axes)
  geom_point() #marks that represent the data values
```
You can do the same with qplot(), which is short for quick plot...

```{r qplot}
qplot(x = displ, y = hwy, data = df, geom = "point")
```

**Challenge 1**
What does "displ" mean? What does "hwy" mean? What is the relationship between these variables? 

Some geoms summarize data or provide trends. For example, <code>geom_smooth()</code> plots the relationship between two variables as smoothed conditional means.

```{r smooth}
df %>%
  ggplot(aes(displ, hwy)) + 
  geom_point() + 
  geom_smooth()
```
Remember, this is equivalent to the example written above...
```{r equivalent}
ggplot(data = df, mapping = aes(x = displ, y = hwy)) + 
  geom_smooth() + geom_point()
```
We can get some useful information out of this plot, like cars that don't follow the trend.

```{r outlier}
ggplot(data = df, mapping = aes(x = displ, y = hwy)) + 
  geom_point(aes(color = displ > 5 & hwy > 20)) + geom_smooth()
```
These cars don't follow the trend, getting higher mpg with bigger engines. We can identify these cars in the data.frame...

**Challenge 2**
Get the cars that were identified in blue and print the data.frame (tibble)
```{r tibble}
View(df %>%
  filter(displ > 5 & hwy > 20))
```

All of these cars have an "fl" of "p". What does that mean? 
How would you find out?
```{r}
unique(df$fl)

ggplot(mpg, aes(x=fl, y=hwy)) + geom_boxplot() + facet_wrap(~cyl, nrow=1)
```

You can add a third variable, like class, to a two dimensional scatterplot by mapping it to an aesthetic. An aesthetic is a visual property of the objects in your plot.Aesthetics include things like the size, the shape, or the color of your points. You can display a point in 
different ways by changing the values of its aesthetic properties.

```{r coloring}
ggplot(data = df) + 
  geom_point(mapping = aes(x = displ, y = hwy, color = class))
```
Notice that a legend is automatically created.
Grouping variables, such as color, might include:

position (i.e., on the x and y axes)
color (“outside” color)
fill (“inside” color)
shape (of points)
linetype
size

## Linetypes

Here are the linetypes
```{r linetypes}
generateRLineTypes<-function(){
  oldPar<-par()
  par(font=2, mar=c(0,0,0,0))
  plot(1, pch="", ylim=c(0,6), xlim=c(0,0.7),  axes=FALSE,xlab="", ylab="")
  for(i in 0:6) lines(c(0.3,0.7), c(i,i), lty=i, lwd=3)
  text(rep(0.1,6), 0:6, labels=c("0.'blank'", "1.'solid'", "2.'dashed'", "3.'dotted'",
                                 "4.'dotdash'", "5.'longdash'", "6.'twodash'"))
  par(mar=oldPar$mar,font=oldPar$font )
}
generateRLineTypes()
```

## Symbols

Here's an example of symbols and their numbers
```{r symbols}
d <- data.frame(p=c(0:25,32:127))
ggplot() +
  scale_y_continuous(name="") +
  scale_x_continuous(name="") +
  scale_shape_identity() +
  geom_point(data=d, mapping=aes(x=p%%16, y=p%/%16, shape=p), size=5, fill="red") +
  geom_text(data=d, mapping=aes(x=p%%16, y=p%/%16+0.25, label=p), size=3)
```
There are 32 quickly referenced geoms, including:

geom_abline   
geom_blank    
geom_errorbar   
geom_errorbarh    
geom_hline    
geom_jitter   
geom_linerange    
geom_point    
geom_pointrange   
geom_polygon    
geom_rect   
geom_rug    
geom_segment    
geom_step   
geom_text   
geom_vline    

See them all here: http://sape.inf.usi.ch/quick-reference/ggplot2/geom

Try them now.
```{r linetype}
df %>%
  ggplot(aes(x = displ, y = hwy)) + 
  geom_line(aes(linetype = class)) + geom_point(aes(color = class))
```

**Challenge 3**
What happens if you move the grouping aesthetic "color = class" to the ggplot() function? Why?    

Try it:
```{r, eval=FALSE}
df %>%
  ggplot(...) + geom_line() + geom_point()
```


## Play time
Keep in mind which geoms you would like to apply to your data.
```{r play}
load("data/eyedata.RData")

# View(df)

df %>%
  mutate(dif = Stop - Start) %>%
  group_by(Participant) %>%
  ggplot(aes(x = Participant, y = dif, group = Participant)) + 
  geom_boxplot()

df <- mpg

df %>%
  ggplot(aes(x = displ, y = hwy)) + geom_bin2d()

df %>% 
  ggplot(aes(x = manufacturer)) + 
  stat_count() + 
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
```

If you really want to create some ridiculous plots...
```{r themes}
#install.packages("ggthemes")
#install.packages("xkcd")
require(ggthemes)
require(xkcd)

if( 'xkcd' %in% fonts()) {
p <- ggplot() + geom_point(aes(x=mpg, y=wt), data=mtcars) +
  theme(text = element_text(size = 16, family = "xkcd"))
} else {
  warning("No xkcd fonts installed!")
  p <- ggplot() + geom_point(aes(x=mpg, y=wt), data=mtcars)
  }
p
### How to get it (I wouldn't mess with this unless you know about installing fonts) ###
# download.file("http://simonsoftware.se/other/xkcd.ttf",dest="xkcd.ttf", mode="wb")
# system("mkdir ~/.fonts")
# system("cp ~/.fonts/xkcd.ttf")
# font_import(pattern = "xkcd", prompt=FALSE)
# fonts()
# fonttable()
# 
# if(.Platform$OS.type != "unix") {
#    ## Register fonts for Windows bitmap output
#      loadfonts(device="pdf")
#    } else {
#     loadfonts()
#     }


#### All of the code below should work unless it's xkcd related and you haven't
#### installed the fonts.

df %>%
  ggplot(aes(x = manufacturer)) + 
  stat_count() + theme_fivethirtyeight() + #fivethirtyeight theme
  labs(title = "Look at me \n I'm Nate Silver") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1),
        plot.title = element_text(hjust = .5))


df %>% 
  ggplot(aes(x = manufacturer)) + 
  stat_count(fill = "#D4E8E7") + theme_tufte() +
  labs(title = "...or Edward Tufte") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1),
        plot.title = element_text(hjust = .5))


df %>% 
  ggplot(aes(x = manufacturer)) + 
  stat_count(position = position_jitter(h = 0.0005), fill = "grey") + 
  theme_xkcd() +
  labs(title = "Now I'm Randall Munroe") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1),
        plot.title = element_text(hjust = .5))

df %>%
  ggplot(aes(x = displ, y = hwy, color = class)) +
  theme_xkcd() + 
  labs(title = "Because who needs \n to interpret your plots?") +
  geom_point(aes(size = cyl),
             alpha = .5) + 
  theme(plot.title = element_text(hjust = .5))
```
Now for serious plots



**Challenge 4**
```{r, darkgreen}
ggplot(data = df) + 
  geom_point(mapping = aes(x = displ, y = hwy, color = "blue"))
```
Why are the dots not blue?


facet_wrap can create multiple plots based on levels of a factor (column)...
```{r wrapping}
ggplot(data = df) + 
  geom_point(mapping = aes(x = displ, y = hwy)) + 
  facet_wrap(~ class, nrow = 2)
```
facet_grid takes slightly different arguments, and facets the plot on two variables...
```{r grid}
ggplot(data = df) + 
  geom_point(mapping = aes(x = displ, y = hwy)) + 
  facet_grid(drv ~ cyl)
```
If you prefer to not facet in the rows or columns dimension, use a . instead of a variable name, e.g. + facet_grid(. ~ cyl).

...facet on a continuous variable?
```{r wrapcont}
ggplot(data = df) + 
  geom_point(mapping = aes(x = class, y = hwy)) + 
  facet_wrap(~cty)
```


**Challenge 5**   
What plots does the following code make? What does . do?
```{r challenge5}
ggplot(data = df) + 
  geom_point(mapping = aes(x = displ, y = hwy)) +
  facet_grid(drv ~ .)
```
hint: try switching the . and drv


You can group by color, shape, linetype, etc. or facet, but you can also apply geoms to only a subset of the data. Below, our smooth line displays just a subset of the mpg dataset, the subcompact cars. The local data argument in geom_smooth() overrides the global data argument in ggplot() for that layer only.

```{r se}
ggplot(data = df, mapping = aes(x = displ, y = hwy)) + 
  geom_point(mapping = aes(color = class)) + 
  geom_smooth(data = filter(mpg, class == "pickup"), se = FALSE)
```
Notice how we used the filter function in the geom function?


## Statistical Transformations
Let's get a new dataset:
```{r diamonds}
df <- ggplot2::diamonds
```
Look in the usual places for information on your diamonds datset.
```{r daimonddim}
ggplot(data = df) + 
  geom_bar(mapping = aes(x = cut)) + 
  theme_classic()
```
In this case, the geom_bar function has a "count" statistic. <code>count</code> is the default statistic for <code>geom_bar()</code> You can change it
```{r dens}
ggplot(data = df) + 
  geom_bar(mapping = aes(x = cut),stat = "density") + 
  theme_classic()
```
...but it doesn't make sense because we have a categorical variable on the x-axis. Remember...
```{r eval=FALSE}
?geom_bar
```
We can also do the counting beforehand and then plot with the "identity" statistic...
```{r counting}
df %>%
  group_by(cut) %>%
  count() 
```
```{r classic}
df %>%
  group_by(cut) %>%
  count() %>%
  ggplot() + geom_bar(mapping = aes(x = cut, y = n),stat = "identity") +
  theme_classic()
```
Now we had to specify the y value because that is how high the bar should be.

We can also use geom_point with stat = "identity", which usually creates scatterplots

```{r cutgroup}
df %>%
  group_by(cut) %>%
  count() %>%
  ggplot() + geom_point(mapping = aes(x = cut, y = n),stat = "identity") +
  theme_classic()
```
Here's a cool trick, we can use proportion in the argument to get fractions of the whole:
```{r group1}
ggplot(data = df) + 
  geom_bar(mapping = aes(x = cut, y = ..prop.., group = 1))
```
Notice that we had to specify the computed variable in dots .."stat"..
```{r ycount}
ggplot(data = df) + 
  geom_bar(mapping = aes(x = cut, y = ..count.., group = 1))
```
also you have to specify that the proportion is out of a group of 1 and not cut. Otherwise they would all be a proportion of 1 for their group.

We can do this before in the data.frame as well:
```{r sum}
df %>%
  group_by(cut) %>%
  count() %>%
  ungroup() %>%
  mutate(prop = n/sum(n)) %>%
  ggplot() + geom_bar(aes(x = cut, y = prop), stat = "identity")
```
You can get summary stats inside ggplot with stat_summary:

```{r allstats}
ggplot(data = df) + 
  stat_summary(
    mapping = aes(x = cut, y = depth),
    fun.ymin = min,
    fun.ymax = max,
    fun.y = median
  )
```

## Changing the position
We can also fill, stack, dodge and jitter our geoms:
```{r positions}
ggplot(data = df) + 
  geom_bar(mapping = aes(x = cut, colour = cut)) #color is the outside
ggplot(data = df) + 
  geom_bar(mapping = aes(x = cut, fill = cut)) #fill is the inside

ggplot(data = df) + 
  geom_bar(mapping = aes(x = cut, fill = clarity)) #fill by a group NOT on the x axis
```
Remember colorBrewer?
```{r brewer}
require(RColorBrewer)
ggplot(data = df) + 
  geom_bar(mapping = aes(x = cut, fill = clarity)) + 
  scale_fill_brewer(palette = 4) + theme_bw()
```
The stacking is performed automatically by the position adjustment specified by the position argument. If you don’t want a stacked bar chart, you can use one of three other options: "identity", "dodge" or "fill".

position = "identity" will place each object exactly where it falls in the context of the graph. 
This is not very useful for bars, because it overlaps them. To see that overlapping we either need to make the bars slightly transparent by setting alpha to a small value, or completely transparent by setting fill = NA.
```{r bars}
ggplot(data = df, mapping = aes(x = cut, fill = clarity)) + 
  geom_bar(alpha = 1/5, position = "identity")
ggplot(data = df, mapping = aes(x = cut, colour = clarity)) + 
  geom_bar(fill = NA, position = "identity") + theme_classic()
```
The identity position adjustment is more useful for 2d geoms, like points, where it is the default.

position = "fill" works like stacking, but makes each set of stacked bars the same height. 
This makes it easier to compare proportions across groups.
```{r masbars}
ggplot(data = df) + 
  geom_bar(mapping = aes(x = cut, fill = clarity), position = "fill")
```
position = "dodge" places overlapping objects directly beside one another. 
This makes it easier to compare individual values.
```{r jitter}
ggplot(data = df) + 
  geom_bar(mapping = aes(x = cut, fill = clarity), position = "dodge") +
  theme_minimal()

ggplot(data = df, aes(x = price, y = carat)) +
  geom_point(position = "jitter") + theme_light() #position jitter adds randomness
```
There are packages that deal with overplotting, but some geoms make this easy as well.
```{r bin2d}
mprice <- df %>%
  group_by(cut,round(carat, 0)) %>%
  summarise(mprice = mean(price))

ggplot(data = df, aes(x = price, y = carat)) +
  geom_bin2d() 

ggplot(data = df, aes(x = price, y = carat)) +
  geom_density2d() + theme_classic()
``` 

## Coordinates
```{r coords}
ggplot(df, aes(x = cut, y = depth)) + geom_boxplot()

ggplot(df, aes(x = cut, y = depth)) + 
  geom_boxplot() +
  coord_flip()
```

### Plot maps if you want
```{r maps}
# install.packages("maps")
require(maps)
globe <- map_data(map = "world")

ggplot(globe, aes(long, lat, group = group)) +
  geom_polygon(fill = "white", colour = "black")

unique(globe$region)[1:10]

newplaces <- map_data(map = "world", region = "Vanuatu")

ggplot(newplaces, aes(long, lat, group = group)) +
  geom_polygon(fill = "white", colour = "black") +
  coord_quickmap() #This adds lat and long
```
Plot with flipped axes or in polar coordinates...
```{r polar}
bar <- ggplot(data = df) + 
  geom_bar(
    mapping = aes(x = cut, fill = cut), 
    show.legend = FALSE,
    width = 1
  ) + 
  theme(aspect.ratio = 1) +
  labs(x = NULL, y = NULL)

bar + coord_flip()
bar + coord_polar()
```
So much you can do!
